search before moving into its component
